abstracted filtering logic away from component
